ml/requirements.txt
pandas==1.5.3
numpy==1.25.2
fuzzywuzzy==0.18.0
sqlalchemy==1.4.52
python-dotenv==1.0.1